Update while.adoc
None of the While or While Loop explanations or tutorials make it entirely clear how the loop ends or what happens when it does end.
Also, NONE of these explanations or tutorials mentions that you can use <break> to get out of them early.   An example using break would be good on this page, too.